requirements.txt
asttokens==2.0.5
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.1
idna==3.3
ipykernel==6.9.1
ipython==8.1.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
keyboard==0.13.5
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
MouseInfo==0.1.3
mss==6.1.0
multitasking==0.0.10
nest-asyncio==1.5.4
numpy==1.22.2
opencv-python==4.5.5.64
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.28
pure-eval==0.2.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.11.2
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.1.0
yfinance==0.1.70